refactor: rename settings -> admin
