+++ crates/client/src/pages/mod.rs
@@ -7,8 +7,8 @@ pub use plugin_manager::*;
 mod search;
 pub use search::*;
 
-mod settings;
-pub use settings::*;
+mod admin;
+pub use admin::*;
 
 mod stats;
 pub use stats::*;
